Modified Desktop part - improved version
