+++ BankAppDesktop/Views/Converters/BankAccountDisplayConverter.cs
@@ -1,18 +1,20 @@
 ﻿using System;
-using Microsoft.UI.Xaml;
+using LoanShark.Domain;
 using Microsoft.UI.Xaml.Data;
 
-namespace LoanShark.Helper
+namespace BankAppDesktop.Views.Converters
 {
-    public class ListNonEmptyToVisibilityConverter : IValueConverter
+    public class BankAccountDisplayConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            if (value is int count)
+            if (value is BankAccount account)
             {
-                return count != 0 ? Visibility.Visible : Visibility.Collapsed;
+                return string.IsNullOrEmpty(account.Name)
+                    ? $"{account.Iban} ({account.Currency})"
+                    : account.Name;
             }
-            return Visibility.Collapsed;
+            return "Unknown Account";
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
