+++ BankAppDesktop/Views/Pages/BankAccountListView.xaml
@@ -4,7 +4,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using: BankAppDesktop.Views"
-    xmlns:conv="using: BankAppDesktop.Converters"
+    xmlns:conv="using: BankAppDesktop.Views.Converters"
     xmlns:dom="using:Common.Models.Bank"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
